+++ cotton-usb-host/src/usb_bus.rs
@@ -16,8 +16,8 @@ use futures::future::FutureExt;
 use futures::{Future, Stream, StreamExt};
 
 pub use crate::host_controller::{
-    DataPhase, DeviceStatus, HostController, InterruptPacket, TransferType,
-    UsbError, UsbSpeed,
+    DataPhase, DeviceStatus, HostController, InterruptPacket, TransferExtras,
+    TransferType, UsbError, UsbSpeed,
 };
 
 /// Basic information about a USB device, perhaps sufficient to select a driver
@@ -369,6 +369,7 @@ impl<HC: HostController> HubState<HC> {
             if p.is_none() {
                 *p = Some(hc.try_alloc_interrupt_pipe(
                     address,
+                    TransferExtras::Normal,
                     endpoint,
                     max_packet_size as u16,
                     interval_ms,
@@ -417,12 +418,16 @@ impl<HC: HostController> Stream for HubStateStream<'_, HC> {
 ///
 pub struct UsbBus<HC: HostController> {
     driver: HC,
+    root_device_speed: Cell<Option<UsbSpeed>>,
 }
 
 impl<HC: HostController> UsbBus<HC> {
     /// Create a new USB host bus from a host-controller driver
     pub fn new(driver: HC) -> Self {
-        Self { driver }
+        Self {
+            driver,
+            root_device_speed: Cell::new(None),
+        }
     }
 
     /// Obtain a stream of hotplug/hot-unplug events
@@ -489,12 +494,7 @@ impl<HC: HostController> UsbBus<HC> {
 
         futures::stream::select(
             root_device.map(InternalEvent::Root),
-            HubStateStream { state: hub_state }
-                /*
-                MultiInterruptStream::<HC::MultiInterruptPipe> {
-                    pipe: &hub_state.pipes,
-                } */
-                .map(InternalEvent::Packet),
+            HubStateStream { state: hub_state }.map(InternalEvent::Packet),
         )
         .then(move |ev| {
             let delay_ms = delay_ms_in.clone();
@@ -502,6 +502,7 @@ impl<HC: HostController> UsbBus<HC> {
                 match ev {
                     InternalEvent::Root(status) => {
                         if let DeviceStatus::Present(speed) = status {
+                            self.root_device_speed.set(Some(speed));
                             self.driver.reset_root_port(true);
                             delay_ms(50).await;
                             self.driver.reset_root_port(false);
@@ -619,6 +620,7 @@ impl<HC: HostController> UsbBus<HC> {
             let delay_ms = delay_ms_in.clone();
             async move {
                 if let DeviceStatus::Present(speed) = status {
+                    self.root_device_speed.set(Some(speed));
                     self.driver.reset_root_port(true);
                     delay_ms(50).await;
                     self.driver.reset_root_port(false);
@@ -640,6 +642,23 @@ impl<HC: HostController> UsbBus<HC> {
         })
     }
 
+    /// Are any special transfer settings needed?
+    ///
+    /// We need to add a preamble if talking to a Low-Speed device via
+    /// a Full-Speed hub (see USB 2.0 section 8.6.5); note that we
+    /// don't need to traverse the whole bus for this, as if the root
+    /// device is FS and our target is LS then there *must* be a hub
+    /// involved somewhere.
+    fn get_transfer_extras(&self, speed: UsbSpeed) -> TransferExtras {
+        if self.root_device_speed.get() == Some(UsbSpeed::Full12)
+            && speed == UsbSpeed::Low1_5
+        {
+            TransferExtras::WithPreamble
+        } else {
+            TransferExtras::Normal
+        }
+    }
+
     /// Configures a device, moving it from "Address" to "Configured" state
     ///
     /// See USB 2.0 figure 9-1. "Configured" state is the useful one, where
@@ -649,9 +668,11 @@ impl<HC: HostController> UsbBus<HC> {
         device: UnconfiguredDevice,
         configuration_value: u8,
     ) -> Result<UsbDevice, UsbError> {
+        let transfer_extras = self.get_transfer_extras(device.usb_speed);
         self.driver
             .control_transfer(
                 device.address(),
+                transfer_extras,
                 device.packet_size_ep0,
                 SetupPacket {
                     bmRequestType: HOST_TO_DEVICE,
@@ -678,12 +699,14 @@ impl<HC: HostController> UsbBus<HC> {
         &self,
         speed: UsbSpeed,
     ) -> Result<(UnaddressedDevice, DeviceInfo), UsbError> {
+        let transfer_extras = self.get_transfer_extras(speed);
         // Read prefix of device descriptor
         let mut descriptors = [0u8; 18];
         let sz = self
             .driver
             .control_transfer(
                 0,
+                transfer_extras,
                 8,
                 SetupPacket {
                     bmRequestType: DEVICE_TO_HOST,
@@ -707,6 +730,7 @@ impl<HC: HostController> UsbBus<HC> {
             .driver
             .control_transfer(
                 0,
+                transfer_extras,
                 packet_size_ep0,
                 SetupPacket {
                     bmRequestType: DEVICE_TO_HOST,
@@ -745,9 +769,11 @@ impl<HC: HostController> UsbBus<HC> {
         device: UnaddressedDevice,
         address: u8,
     ) -> Result<UnconfiguredDevice, UsbError> {
+        let transfer_extras = self.get_transfer_extras(device.usb_speed);
         self.driver
             .control_transfer(
                 0,
+                transfer_extras,
                 device.packet_size_ep0,
                 SetupPacket {
                     bmRequestType: HOST_TO_DEVICE,
@@ -805,9 +831,11 @@ impl<HC: HostController> UsbBus<HC> {
         setup: SetupPacket,
         data_phase: DataPhase<'_>,
     ) -> Result<usize, UsbError> {
+        let transfer_extras = self.get_transfer_extras(device.usb_speed);
         self.driver
             .control_transfer(
                 device.usb_address,
+                transfer_extras,
                 device.packet_size_ep0,
                 setup,
                 data_phase,
@@ -826,6 +854,7 @@ impl<HC: HostController> UsbBus<HC> {
         self.driver
             .control_transfer(
                 ep.usb_address,
+                TransferExtras::Normal,
                 8,
                 SetupPacket {
                     bmRequestType: 2,
@@ -905,14 +934,16 @@ impl<HC: HostController> UsbBus<HC> {
     ///  - interval_ms: polling interval, in milliseconds
     pub fn interrupt_endpoint_in(
         &self,
-        address: u8,
+        device: &UsbDevice,
         endpoint: u8,
         max_packet_size: u16,
         interval_ms: u8,
     ) -> impl Stream<Item = InterruptPacket> + '_ {
+        let transfer_extras = self.get_transfer_extras(device.usb_speed);
         self.driver
             .alloc_interrupt_pipe(
-                address,
+                device.usb_address,
+                transfer_extras,
                 endpoint,
                 max_packet_size,
                 interval_ms,
@@ -936,6 +967,7 @@ impl<HC: HostController> UsbBus<HC> {
         device: &UnconfiguredDevice,
         visitor: &mut impl DescriptorVisitor,
     ) -> Result<(), UsbError> {
+        let transfer_extras = self.get_transfer_extras(device.usb_speed);
         // Handle descriptor suites > 64 byte
         //   256 is arbitrary, but should be enough for most devices
         //   TinyUSB uses 256 - https://github.com/hathach/tinyusb/blob/5572168994a29266df6cbf12b46919498d3ece66/examples/host/hid_controller/src/tusb_config.h#L103
@@ -947,6 +979,7 @@ impl<HC: HostController> UsbBus<HC> {
             .driver
             .control_transfer(
                 device.address(),
+                transfer_extras,
                 device.packet_size_ep0,
                 SetupPacket {
                     bmRequestType: DEVICE_TO_HOST,
@@ -967,6 +1000,7 @@ impl<HC: HostController> UsbBus<HC> {
                     .driver
                     .control_transfer(
                         device.address(),
+                        transfer_extras,
                         device.packet_size_ep0,
                         SetupPacket {
                             bmRequestType: DEVICE_TO_HOST,
@@ -1027,6 +1061,7 @@ impl<HC: HostController> UsbBus<HC> {
             .driver
             .control_transfer(
                 device.address(),
+                TransferExtras::Normal,
                 device.packet_size_ep0,
                 SetupPacket {
                     bmRequestType: DEVICE_TO_HOST | CLASS_REQUEST,
@@ -1064,6 +1099,7 @@ impl<HC: HostController> UsbBus<HC> {
         self.driver
             .control_transfer(
                 hub_address,
+                TransferExtras::Normal,
                 8,
                 SetupPacket {
                     bmRequestType: DEVICE_TO_HOST
@@ -1095,6 +1131,7 @@ impl<HC: HostController> UsbBus<HC> {
         self.driver
             .control_transfer(
                 hub_address,
+                TransferExtras::Normal,
                 8,
                 SetupPacket {
                     bmRequestType: HOST_TO_DEVICE
@@ -1120,6 +1157,7 @@ impl<HC: HostController> UsbBus<HC> {
         self.driver
             .control_transfer(
                 hub_address,
+                TransferExtras::Normal,
                 8,
                 SetupPacket {
                     bmRequestType: HOST_TO_DEVICE